+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Bump clj-kondo to `2024.05.22` fixing high memory usage issue.
   - Bump Clojure to `1.11.3`
   - Fix use of `:filename-pattern` from kondo breaking references.
+  - Add `api.dump` to the pod namespace so it can be used with babashka.
 
 - Editor
   - Add `clojure.test/deftest` option to resolve macro as code action. #1850